Move utils to utils to expose for testing purposes
