+++ packages/format-library/src/link/inline.js
@@ -24,112 +24,8 @@ import { useSelect } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import { createLinkFormat, isValidHref } from './utils';
+import { createLinkFormat, isValidHref, getFormatBoundary } from './utils';
 import { link as settings } from './index';
-/**
- * External dependencies
- */
-import { find } from 'lodash';
-
-function getFormatBoundary(
-	value,
-	format,
-	startIndex = value.start,
-	endIndex = value.end
-) {
-	const EMPTY_BOUNDARIES = {
-		start: null,
-		end: null,
-	};
-
-	const { formats } = value;
-	const newFormats = formats.slice();
-
-	// If there are no matching formats *at* the reputed "end"
-	// then we are already at the end of the format.
-	// This is because endIndex is always +1 more than the
-	// end of the format itself.
-	const isIndexAtEnd = ! find( newFormats[ endIndex ], {
-		type: format.type,
-	} );
-
-	// Account for endindex being +1 (see above).
-	const initialIndex = isIndexAtEnd ? startIndex - 1 : startIndex;
-
-	// Retrieve a reference to the target format object
-	// at the initial index.
-	const targetFormat = getTargetFormatObjectReference(
-		newFormats,
-		initialIndex,
-		format.type
-	);
-
-	if ( ! targetFormat ) {
-		return EMPTY_BOUNDARIES;
-	}
-
-	const index = newFormats[ initialIndex ].indexOf( targetFormat );
-
-	// If not found...
-	if ( index === -1 ) {
-		return EMPTY_BOUNDARIES;
-	}
-
-	// Walk the startIndex "backwards" to the leading "edge" of the matching format.
-	startIndex = walkToBoundary(
-		newFormats,
-		initialIndex,
-		targetFormat,
-		index,
-		'backwards'
-	);
-
-	// Walk the endIndex "forwards" until the trailing "edge" of the matching format.
-	endIndex = walkToBoundary(
-		newFormats,
-		initialIndex,
-		targetFormat,
-		index,
-		'forwards'
-	);
-
-	// Safe guard: start index cannot be less than 0
-	startIndex = startIndex < 0 ? 0 : startIndex;
-
-	// // Return the indicies of the "edges" as the boundaries.
-	return {
-		start: startIndex,
-		end: endIndex,
-	};
-}
-
-// Retrieve a reference to the target format object
-// at the initial index.
-function getTargetFormatObjectReference( formats, index, type ) {
-	return find( formats[ index ], {
-		type,
-	} );
-}
-
-function walkToBoundary( formats, index, initFormat, formatIndex, direction ) {
-	while (
-		formats[ index ] &&
-		formats[ index ][ formatIndex ] === initFormat
-	) {
-		if ( direction === 'forwards' ) {
-			index++;
-		} else {
-			index--;
-		}
-	}
-
-	// Restore by one to avoid out of bounds.
-	if ( direction === 'backwards' ) {
-		index++;
-	}
-
-	return index;
-}
 
 function InlineLinkUI( {
 	isActive,
